Added a new page and JS application for start location and destination
Since the main JS file(travelshed.application.js) was throwing errors
when the user clicked the search button, I added a new
page(schoolRoutes.html) and JS application(schoolroutes.application.js)
based on the scenic template. The template has support for the start and
end locations out of the box.